Change `WatchEvent::Bookmark` to store the resource version directly instead of the full object.
In some cases, the API server sends a BOOKMARK watch event that can't be
successfully deserialized as the resource type. The example in #70 is of
a watch event for pods where the `spec.containers` field is `null`, even though
the spec requires that `containers` be non-null.

Since the only point of bookmark events is to carry a resource version,
this change makes it so that the deserializer only looks for
the `metadata.resourceVersion` field, in addition to the `apiVersion`
and `kind` fields like before.

Apart from changing the definition of the `WatchEvent::Bookmark` variant,
this change also adds a bound to the type parameter of `WatchEvent` -
it must now also implement `k8s_openapi::Resource` since its API_VERSION
and KIND are still needed to validate the bookmark object.

Fixes #70